DAY-1/Fake Product Review Detection/ml-model/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
joblib==1.3.2
textblob==0.17.1
gunicorn==21.2.0
python-dotenv==1.0.0
